Translation_Orchestrator_Agent/pyproject.toml
[project]
name = "translation-orchestrator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "generativeai>=0.0.1",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
]
